+++ www/SavedState.cpp
@@ -1,7 +1,7 @@
 #include "SavedState.h"
 
 
-void SavedState::save(glm::ivec2 newSavedPlayerPosition, bool newSavedUpsideDown, int newSavedScreen)
+void SavedState::save(glm::vec2 newSavedPlayerPosition, bool newSavedUpsideDown, int newSavedScreen)
 {
 	this->savedPlayerPosition = newSavedPlayerPosition;
 	this->savedUpsideDown = newSavedUpsideDown;
